% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project_audit.R
\name{check_data_lvl_1}
\alias{check_data_lvl_1}
\title{Check the top level data directory}
\usage{
check_data_lvl_1(proj_dir, ignore_patterns = c("^\\\\d{2}", "^readme"),
  pattern_ignore_case = TRUE, verbose = FALSE)
}
\arguments{
\item{proj_dir:}{root project directory}

\item{ignore_patterns:}{vector of regular expressions to ignore}

\item{pattern_ignore_case:}{boolean on wether to ignore case in regular expression match}

\item{verbose:}{boolean on whether or not to print statements}
}
\value{
list of values
}
\description{
Check the top level data directory
}
