% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project_audit.R
\name{check_proj_top_level}
\alias{check_proj_top_level}
\title{Checks the root project folder structure}
\usage{
check_proj_top_level(proj_dir, should_contain_exactly = c(".git",
  ".gitignore", "data", "output", "src"),
  should_contain_pattern = c("*.rproj$", "^readme*"), verbose = FALSE)
}
\arguments{
\item{proj_dir:}{root project directory}

\item{should_contain_exactly:}{files or folders that should exist here}

\item{should_contain_pattern:}{regex pattern of files or folders that should exist here}

\item{verbose:}{boolean whether statements should be printed}
}
\value{
list of values
}
\description{
Checks the root project folder structure
}
