% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project_audit.R
\name{check_src_folder}
\alias{check_src_folder}
\title{Check source folder for abnormal files}
\usage{
check_src_folder(proj_dir, r_files_patterns = c("*.r$", "*.rmd$"),
  ignore_patterns = c(".gitignore", "readme*", "*.py"),
  data_patterns = c("*.csv$", "*.xls$", "*.xlsx$", "*.rdata$"),
  pattern_ignore_case = TRUE, verbose = FALSE)
}
\arguments{
\item{proj_dir:}{root project directory}

\item{r_files_patterns:}{vector of regular expressions to match R script files}

\item{ignore_patterns:}{vector of regular expressions of files to ignore in the source folder}

\item{data_patterns:}{vector of regular expressions to look for data files that should not exist in a source folder}

\item{pattern_ignore_case:}{bool on whether or not to ignore case when doing regex match}

\item{verbose:}{bool on whether to print statements}
}
\value{
list of values
}
\description{
Check source folder for abnormal files
}
