% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project_audit.R
\name{check_wd}
\alias{check_wd}
\title{Check the root working directory}
\usage{
check_wd(base_dir = getwd(), verbose = FALSE)
}
\arguments{
\item{base_dir:}{base directory to check working directories}

\item{verbose:}{boolean on whether to print statements}
}
\value{
list of values
}
\description{
Check the root working directory
}
