% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/con_db.R
\name{con_db_postgresql}
\alias{con_db_postgresql}
\title{Makes a connection to a postgresql database}
\usage{
con_db_postgresql(dbname, user, pass, host, port, close_existing_cons)
}
\value{
postgresql database connection
}
\description{
Makes a connection to a postgresql database
}
