% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_vault_sdal.R
\name{create_vault_sdal}
\alias{create_vault_sdal}
\title{Create the default sdal vault}
\usage{
create_vault_sdal(vault = "/home/sdal/projects/sdal/vault/")
}
\description{
mainly useful for testing or after deleting the vault
}
