% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_acs_key.R
\name{get_acs_key}
\alias{get_acs_key}
\title{Retrieves the ACS API key on the server}
\usage{
get_acs_key(file = "/home/sdal/projects/keys/acs_api_key.txt")
}
\arguments{
\item{file}{location of the google maps key file by default it points to the location in the sdal server}
}
\value{
string
}
\description{
So it's not accidentently pushed in the code
}
\examples{
\dontrun{
get_acs_key()

get_acs_key('~/acs_api_key.txt')
}
}
