% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project_audit.R
\name{get_all_files}
\alias{get_all_files}
\title{Recursively get all files in a given path}
\usage{
get_all_files(proj_dir)
}
\arguments{
\item{proj_dir:}{root path to start recursive search}
}
\value{
character vector of files
}
\description{
Recursively get all files in a given path
}
