% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_code_gov_key.R
\name{get_code_gov_key}
\alias{get_code_gov_key}
\title{Retrieves the code.gov key on the server}
\usage{
get_code_gov_key(file = "/home/sdal/projects/sdal/api_keys/code_gov.txt")
}
\arguments{
\item{file}{location of the google maps key file by default it points to the location in the sdal server}
}
\value{
string
}
\description{
So it's not accidentently pushed in the code
}
\examples{
\dontrun{
get_code_gov_key()

get_code_gov_key('~/google_maps_api_key.txt')
}
}
