% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project_audit.R
\name{get_dir_files}
\alias{get_dir_files}
\title{Non-recursively get files in a give path}
\usage{
get_dir_files(folder)
}
\arguments{
\item{folder:}{root path to look for files}
}
\value{
character vector of files
}
\description{
Non-recursively get files in a give path
}
