% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_google_maps_key.R, R/radar_search.R
\name{get_google_maps_key}
\alias{get_google_maps_key}
\title{Retrieves the Google Maps API key on the server}
\usage{

  get_google_maps_key(file = "/home/sdal/projects/keys/google_maps_api_key.txt")


  get_google_maps_key(file = "/home/sdal/projects/keys/google_maps_api_key.txt")
}
\arguments{
\item{file}{location of the google maps key file by default it points to the location in the sdal server}

\item{lat}{the latitude}

\item{lon}{the longitude}

\item{type_search_string}{the string to be used to look for location types}

\item{radius}{how wide the radar search will look}

\item{...}{optional. parameters passed into sdal::get_google_maps_key}
}
\description{
So it's not accidentently pushed in the code

Uses the google maps api which will require an API key
}

