% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ip_address.R
\name{get_ip_address}
\alias{get_ip_address}
\title{Get the current IP address}
\usage{
get_ip_address(cmd = "ip address",
  ip_pattern = "(\\\\d{1,3}\\\\.){3}\\\\d{1,3}", con_pattern = "eth0",
  verbose = FALSE)
}
\arguments{
\item{cmd}{system call that gets the IP address. Defaults to `ip address`}

\item{ip_pattern}{regex pattern that matches an ip address}

\item{con_pattern}{regex pattern for the device. Defaults to 'eth0'}

\item{verbose}{whether or not to print statments during parsing. Defaults to FALSE}
}
\value{
string
}
\description{
An example use cause is
getting the IP address within a docker container
for the user to SSH into in order to run an R script in the background
}
\examples{
get_ip_address()
get_ip_address(con_pattern = 'wlp3s0')
}
