% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user_pass.R
\name{get_my_username}
\alias{get_my_username}
\title{Get the username of the current user}
\usage{
get_my_username(secret_name = unname(Sys.info()["user"]), key = local_key(),
  vault = "/home/sdal/projects/sdal/vault", verbose = FALSE)
}
\arguments{
\item{secret_name, }{string, the secret to look up, defaults to the current username}

\item{key, }{string, location of the private key, defaults to local_key(), ~/.ssh/id_rsa}

\item{vault, }{string, location of the vault, defualts to /home/sdal/projects/sdal/vault}
}
\value{
character, the username
}
\description{
Get the username of the current user
}
