% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/openhub_api_keys.R
\name{get_openhub_key}
\alias{get_openhub_key}
\title{Gets a single OpenHub API key}
\usage{
get_openhub_key(key_index = 1, key_id = NULL, key_name = NULL,
  file = "/home/sdal/projects/sdal/api_keys/ohloh_keys.csv")
}
\arguments{
\item{key_index}{the row index of the key (can also be vector of integers), defaults to 1}

\item{key_id}{the id to use to subset the table of API keys (use with key_name), defaults to NULL}

\item{key_name}{the name to use to subset the table of API keys (use with key_id), defaults to NULL}

\item{file}{location of the google maps key file by default it points to the location in the sdal server}
}
\value{
string
}
\description{
Calls `get_openhub_keys_df` to get API key(s) so it's not accidentently pushed in the code.
}
\examples{
get_openhub_key()

get_openhub_key(2)

get_openhub_key(c(1, 2))

get_openhub_key(key_id = 'oh_key_dc', key_name = 'Daniel')
}
