% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/openhub_api_keys.R
\name{get_openhub_keys_df}
\alias{get_openhub_keys_df}
\title{Retrieves the OpenHub keys from members in the lab}
\usage{
get_openhub_keys_df(file = "/home/sdal/projects/keys/ohloh_keys.csv")
}
\arguments{
\item{file}{location of the google maps key file by default it points to the location in the sdal server}
}
\value{
dataframe
}
\description{
So it's not accidentently pushed in the code.
}
\details{
Dataframe has 3 columns: id, name, and key.
id is some identifier for the key
name is the name of the person the key is associated with
key is the actuall API key.
}
\examples{
get_openhub_keys_df()
}
