% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/latlong2county.R
\name{latlong2county}
\alias{latlong2county}
\title{Return a county for a data.frame of lat lon values}
\usage{
latlong2county(pointsDF, state_FIPS)
}
\arguments{
\item{pointsdF}{data.frame in which:
column 1 contains the longitude in degrees (negative in the US) AND
column 2 contains the latitude in degrees}
}
\description{
Return a county for a data.frame of lat lon values
}
