% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/latlong2state.R
\name{latlong2state}
\alias{latlong2state}
\title{Return a state for a data.frame of lat lon values}
\usage{
latlong2state(pointsDF)
}
\arguments{
\item{pointsdF}{data.frame in which:
column 1 contains the longitude in degrees (negative in the US) AND
column 2 contains the latitude in degrees}
}
\description{
Return a state for a data.frame of lat lon values
}

