% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/latlongrid.R
\name{latlongrid}
\alias{latlongrid}
\title{Create a grid for a specified state}
\usage{
latlongrid(box_lower_left_lon, box_lower_left_lat, box_upper_right_lon,
  box_upper_right_lat, lon_bearing, lat_bearing, distance, state)
}
\arguments{
\item{box_lower_left_lon}{longitude for lower left corner}

\item{box_lower_left_lat}{latitude for lower left cornet}

\item{box_upper_right_lon}{longitude for upper right corner}

\item{box_upper_right_lat}{latitude for upper right corner}

\item{lon_bearing}{distance to move lon (in feet?)}

\item{lat_bearing}{distance to move lat (in feet?)}

\item{distance}{distance}

\item{state}{state}
}
\description{
Create a grid for a specified state
}

