% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_zipped_shapefile.R
\name{load_zipped_shapefiles}
\alias{load_zipped_shapefiles}
\title{Load all files in a directory into an environment with a given prefix}
\usage{
load_zipped_shapefiles(shapefile_dir, prefix, delete_unzip = TRUE,
  env = .GlobalEnv, verbose = TRUE)
}
\arguments{
\item{shapefile_dir}{path to the directory of files}

\item{prefix}{prefix for the loaded shapefile when it is loaded in the environment}

\item{delete_unzip}{boolean on whether to delete the extracted zip files (default TRUE)}

\item{env}{environment where the file will be loaded into (default to .GlobalEnv)}

\item{verbose}{boolean on whether or not to print the file as they are being loaded (default TRUE)}
}
\description{
Load all files in a directory into an environment with a given prefix
}
\examples{
load_zipped_shapefiles("data/shapefiles/zip", "arl_")

}
