% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiplot.R
\name{multiplot}
\alias{multiplot}
\title{Multiple plot function}
\usage{
multiplot(..., plotlist = NULL, file, cols = 1, layout = NULL)
}
\arguments{
\item{...}{ggplot objects}

\item{plotlist}{if individual ggplot objects are not passed into \code{...},
then a \code{list} of ggplot objects can also be passed}

\item{cols}{number of columns used to layout the plots, defaults to \code{`1`}}

\item{layout}{a matrix specifying the layout. If present, 'cols' is ignored.
If the layout is something like matrix(c(1,2,3,3), nrow=2, byrow=TRUE),
then plot 1 will go in the upper left, 2 will go in the upper right,
and 3 will go all the way across the bottom.}
}
\description{
This function is taken from the R Graphics Coookbook by Winston Chang.
It lays out multiple ggplot objects into a single figure.
}
