% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project_audit.R
\name{project_audit}
\alias{project_audit}
\title{Performs a project template audit}
\usage{
project_audit(project_path = getwd(), verbose = FALSE,
  print_results = FALSE)
}
\arguments{
\item{project_path:}{base directory to start project template audit}

\item{verbose:}{boolean on whether to print statments}

\item{print_results:}{boolean to print_results}
}
\value{
nothing, just prints a report
}
\description{
Performs a project template audit
}
