% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radar_search.R
\name{radar_search}
\alias{radar_search}
\title{Performs a radar search from a given lat and lon within a certain radius}
\usage{
radar_search(lat, lon, type_search_string, radius = 1609, ...)
}
\arguments{
\item{lat}{the latitude}

\item{lon}{the longitude}

\item{type_search_string}{the string to be used to look for location types}

\item{radius}{how wide the radar search will look}

\item{...}{optional. parameters passed into sdal::get_google_maps_key}
}
\value{
list
}
\description{
Uses the google maps api which will require an API key
}
\examples{
sdalr::radar_search(lat = 37.231264, lon = -80.426745,
                    type_search_string = 'grocery_or_supermarket')

sdalr::radar_search(lat = 37.231264, lon = -80.426745,
                    type_search_string = 'grocery_or_supermarket',
                    file = '~/google_maps_api_key.txt')
}

