% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rename_script_numbers.R
\name{rename_script_numbers}
\alias{rename_script_numbers}
\title{Iteratively renames files in a directory using the `renumber_file` function}
\usage{
rename_script_numbers(path, number_match_pattern, pad_width,
  rfile_pattern = "*.[rR]$", dry_run = TRUE, verbose = TRUE)
}
\arguments{
\item{path:}{directory of files}

\item{number_match_pattern:}{pattern passed into `renumber_file`}

\item{pad_width:}{left pad width passed into `renumber_file`}

\item{rfile_pattern:}{regular expression that looks for files to rename,
defaults to looking for r script files}

\item{dry_run:}{whether or not to actually do the rename or not.
defaults to TRUE, which does not rename files,
only prints out pre and post file names}

\item{verbose:}{whether or not to show print statments}
}
\description{
Iteratively renames files in a directory using the `renumber_file` function
}
\examples{
rename_script_numbers()
}
