% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rename_script_numbers.R
\name{renumber_file}
\alias{renumber_file}
\title{Returns a new file name with padded digits}
\usage{
renumber_file(file_name, number_match_pattern, pad_width, verbose = TRUE)
}
\arguments{
\item{file_name:}{original filename to rename}

\item{number_match_pattern:}{regular expression to extract the number to pad}

\item{pad_width:}{total number of digits to left pad with}

\item{verbose:}{print statements}
}
\description{
Returns a new file name with padded digits
}
\examples{
new_filename <- renumber_file(file_name = 'account_table_2017-07-17_1.RData',
                              number_match_pattern = '\\\\d*\\\\.RData$',
                              pad_width = 3,
                              verbose = FALSE)

new_filename <- renumber_file(file_name = "1-my_script.R",
                              number_match_pattern = '^\\\\d*-',
                              pad_width = 3,
                              verbose = FALSE)
}
