% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user_pass.R
\name{setup_user_pass}
\alias{setup_user_pass}
\title{Initial setup of secret vault with username and password secrets}
\usage{
setup_user_pass(username = Sys.info()["user"], password = NULL,
  public_key = "~/.ssh/id_rsa.pub",
  vault = "/home/sdal/projects/sdal/vault",
  secret_name = Sys.info()["user"])
}
\arguments{
\item{username:}{string, the username, defaults to the user of the current R session}

\item{password:}{string, defaults to NULL which will prompt you for your password}

\item{public_key:}{string, the location of the public key, defaults to '~/.ssh/id_rsa.pub'}

\item{vault:}{string, location of the vault, defaults to '/home/sdal/projects/sdal/vault'}

\item{secret_name:}{string, the name of the secret to be stored, defaults to 'ldap'}
}
\description{
If you need to update a username/password please see: `update_user_pass``
}
