% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user_pass.R
\name{update_user_pass}
\alias{update_user_pass}
\title{Update the username and password saved in the vault
Typically used after you change your LDAP password}
\usage{
update_user_pass(username = Sys.info()["user"], password = NULL,
  secret_name = "ldap", vault = "/home/sdal/projects/sdal/vault")
}
\arguments{
\item{username, }{string, username to be updated, defaults to the current user}

\item{password, }{string, password to be updated, defaults to NULl which will prompt the user for a password}

\item{secret_name, }{string, location of the vault, defaults to the sdal vault}
}
\description{
Update the username and password saved in the vault
Typically used after you change your LDAP password
}
