% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zip_matcher.R
\name{zip_matcher}
\alias{zip_matcher}
\title{Filter ACS zipcodes to the ones you want}
\usage{
zip_matcher(tablecode, endyear, ziplist)
}
\arguments{
\item{tablecode}{ACS tablecode. Refers to corresponding demographic variables such as total population. Eg. B01003: Population by Age and Sex}

\item{endyear}{The year of the ACS survey}

\item{ziplist}{A character vector of zipcodes to filter ACS variables by. Can be numeric vector if there are no leading zeros.}
}
\value{
dataframe that returns 3 columns. First column is the zip code from ziplist. Second column is the variable estimate from the table code and end year. Thrid column is the standard error of the second column from the table code and end year.
}
\description{
When you pull the ACS data table by zipcodes, it will provide zipcodes all over the
States. This function allows you to filer to the zipcodes you want.
This requires a ACS key installed for it to work. They function is api.key.install().
Below are the instructions to get ACS API Key.
1. load and install the R package acs.
2. Install an API key by going to the Census website here
   http://api.census.gov/data/key_signup.html
   Fill out the simple form and agree to the Terms of Service.
3. Census will email you a key for your use only.
4. Once you get your key from census, you can install it using api.key.install()
}
\examples{
pop_age_group_2015 <- zip_matcher("B01003",2015, c("22030", "22031"))

}
