% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/con_db.R
\name{con_db}
\alias{con_db}
\title{Makes a connection to a database.}
\usage{
con_db(db, user, driver = "PostgreSQL", host = "localhost", port = 5432,
  close_existing_cons = TRUE)
}
\arguments{
\item{db}{the name of the database you want to connect to}

\item{user}{the username for the database connection}

\item{driver}{the database driver to use, defaults to 'PostgreSQL'}

\item{host}{where the database is running, defaults to 'localhost'}

\item{port}{the port where the database in running, defaults to '5432'}

\item{close_existing_cons}{whether to close existing database connections, defaults to 'TRUE'}

\item{pass}{the password for the database connection}
}
\value{
a database connection
}
\description{
Creates a connection to a database.
This function is really a high-level function that calls the corresponding
\code{con_db} function for a particular database driver.
}
\examples{
con_db('sample_db', 'dan', 'isawesome')
}

