% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_google_maps_key.R
\name{get_google_maps_key}
\alias{get_google_maps_key}
\title{Retrieves the Google Maps API key on the server}
\usage{

  get_google_maps_key(file = "/home/sdal/projects/keys/google_maps_api_key.txt")
}
\arguments{
\item{file}{location of the google maps key file by default it points to the location in the sdal server}
}
\value{
string
}
\description{
So it's not accidentently pushed in the code
}
\examples{
\dontrun{
get_google_maps_key()

get_google_maps_key('~/google_maps_api_key.txt')
}
}
