% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user_pass.R
\name{get_my_password}
\alias{get_my_password}
\title{Get the password of the current user}
\usage{
get_my_password(secret_name = "ldap", key = local_key(),
  vault = "/home/sdal/projects/sdal/vault")
}
\description{
Get the password of the current user
}
