% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user_pass.R
\name{get_my_username}
\alias{get_my_username}
\title{Get the username of the current user}
\usage{
get_my_username(secret_name = "ldap", key = local_key(),
  vault = "/home/sdal/projects/sdal/vault")
}
\description{
Get the username of the current user
}
