% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/latlongrid.R
\name{latlongrid}
\alias{latlongrid}
\title{Create a grid for a specified state}
\usage{
latlongrid(lon_bearing = 90, lat_bearing = 0, distance = 1000, state_fips)
}
\arguments{
\item{lon_bearing}{distance to move lon (in feet?)}

\item{lat_bearing}{distance to move lat (in feet?)}

\item{distance}{distance}

\item{state_fips}{state FIPS code}
}
\description{
Create a grid for a specified state
}
